+++ jacsstorage-core/src/main/java/org/janelia/jacsstorage/coreutils/IOStreamUtils.java
@@ -1,7 +1,5 @@
 package org.janelia.jacsstorage.coreutils;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;