+++ jacsstorage-core/src/main/java/org/janelia/jacsstorage/coreutils/NetUtils.java
@@ -1,14 +1,21 @@
 package org.janelia.jacsstorage.coreutils;
 
+import org.apache.commons.lang3.StringUtils;
+
 import java.net.InetAddress;
 
 public class NetUtils {
+    private static volatile String hostname;
+
     public static String getCurrentHostName() {
-        try {
-            InetAddress ip = InetAddress.getLocalHost();
-            return ip.getHostName();
-        } catch (Exception e) {
-            throw new IllegalStateException(e);
+        if (StringUtils.isBlank(hostname)) {
+            try {
+                InetAddress ip = InetAddress.getLocalHost();
+                hostname = ip.getHostName();
+            } catch (Exception e) {
+                throw new IllegalStateException(e);
+            }
         }
+        return hostname;
     }
 }